requirements.txt
deepspeed==0.9.2
ffmpeg==1.4
more_itertools==9.1.0
numpy==1.23.0
pandas==1.4.2
pytorch_lightning==2.0.2
PyYAML==6.0
PyYAML==6.0
regex==2022.10.31
sacrebleu==2.3.1
torch==1.13.1
torchaudio==0.13.1
torchmetrics==0.11.4
tqdm==4.64.0
transformers==4.29.1
